requirements.txt
backports.csv==1.0.7
certifi==2019.9.11
chardet==3.0.4
CMRESHandler==1.0.0
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.2.5
django-cors-headers==3.1.0
django-crispy-forms==1.7.2
django-filter==2.2.0
django-formtools==2.1
django-haystack==2.8.1
django-import-export==1.2.0
django-qiniu-storage==2.3.1
django-redis==4.10.0
django-reversion==3.0.4
django-smtp-ssl==1.0
django-storages==1.7.2
djangorestframework==3.10.3
djangorestframework-jwt==1.11.0
drf-extensions==0.5.0
drf-haystack==1.8.6
elastic-apm==5.1.2
elasticsearch==7.0.5
et-xmlfile==1.0.1
future==0.15.2
httplib2==0.9.2
idna==2.8
itypes==1.1.0
jdcal==1.4.1
Jinja2==2.10.1
Markdown==3.1.1
MarkupSafe==1.1.1
mysqlclient==1.4.4
oauthlib==3.1.0
odfpy==1.4.0
openpyxl==2.6.3
Pillow==6.1.0
psycopg2==2.8.3
pycryptodome==3.9.0
PyJWT==1.7.1
python-dateutil==2.8.0
python3-openid==3.1.0
pytz==2019.2
PyYAML==5.1.2
qiniu==7.2.6
raven==6.10.0
redis==3.3.8
requests==2.22.0
requests-oauthlib==1.2.0
sentry-sdk==0.13.1
six==1.10.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
sqlparse==0.3.0
tablib==0.13.0
uritemplate==3.0.0
urllib3==1.25.3
xlrd==1.2.0
XlsxWriter==1.2.1
xlwt==1.3.0
